Move testCommand into D.S.Command.Test
